requirements/common.txt
Brotli==1.0.9
Flask-Compress==1.12
Flask==2.2.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Werkzeug==2.2.2
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
dash-bootstrap-components==1.2.1
dash-bootstrap-templates==1.0.6
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
git+https://github.com/sqlalchemy/sqlalchemy@befa8e92675dea992423011b929dfc7ca45de6b6#egg=SQLAlchemy
git+https://github.com/tuokri/rs2simlib@0.4.5#egg=rs2simlib
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
logbook==1.5.3
natsort==8.1.0
numpy>=1.22.4
orjson==3.8.0
pandas==1.4.3
plotly==5.10.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
six==1.16.0
tenacity==8.0.1
urllib3==1.26.11
